refactor tests to no longer use graph_builder
