+++ tests/orchestration/test_node_wiring.py
@@ -2,10 +2,7 @@ import pytest
 from typing import Any
 import pandas as pd
 from pyautocausal.orchestration.nodes import Node, InputNode
-from pyautocausal.orchestration.graph_builder import ExecutableGraph
-from pyautocausal.persistence.output_config import OutputConfig
-from pyautocausal.persistence.output_types import OutputType
-from unittest.mock import Mock
+from pyautocausal.orchestration.graph import ExecutableGraph
 
 
 # Test functions with type hints
