+++ tests/pipelines/test_causal_pipeline.py
@@ -1,7 +1,7 @@
 import pytest
 from pathlib import Path
 import pandas as pd
-from pyautocausal.orchestration.graph_builder import ExecutableGraph
+from pyautocausal.orchestration.graph import ExecutableGraph
 from pyautocausal.pipelines.library import DoubleMLNode, OLSNode
 from pyautocausal.pipelines.example import condition_nObs_DoubleML, condition_nObs_OLS
 from pyautocausal.orchestration.condition import Condition
